🚚 refactor: Rename CloudflareEnv.js to cloudflare-env.js
(cherry picked from commit 1900453915d772afdb2989b86939295e2486470f)
